prepare for more complext stress scenarios
