+++ .envrc
@@ -1,3 +1,3 @@
 function @tasks-3 {
-  METEOR_CHECKOUT_PATH='/Users/leonardo/Repositories/meteor/meteor' ./scripts/monitor.sh tasks-3.x reactive-stress.yml reactive
+  METEOR_CHECKOUT_PATH='/Users/leonardo/Repositories/meteor/meteor' ./scripts/monitor.sh tasks-3.x basic/reactive-stress.yml reactive
 }
