+++ tests/basic/non-reactive.spec.js
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { nonReactiveAddAndRemoveTasks } from './test-helpers';
+import { nonReactiveAddAndRemoveTasks } from '../test-helpers';
 
 test('non-reactive', async ({ page }) => {
   await nonReactiveAddAndRemoveTasks(page);
