+++ tests/basic/reactive.spec.js
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { reactiveAddAndRemoveTasks } from './test-helpers';
+import { reactiveAddAndRemoveTasks } from '../test-helpers';
 
 test('reactive', async ({ page }) => {
   await reactiveAddAndRemoveTasks(page);
